base_agent_tools set for third party agents
