+++ src/crewai/agents/agent_builder/base_agent.py
@@ -1,6 +1,6 @@
 from copy import deepcopy
 import uuid
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional
 from abc import ABC, abstractmethod
 from pydantic import (
     UUID4,
@@ -178,15 +178,6 @@ class BaseAgent(ABC, BaseModel):
     @abstractmethod
     def get_delegation_tools(self, agents: List["BaseAgent"]):
         """Set the task tools that init BaseAgenTools class."""
-        pass
-
-    @abstractmethod
-    def format_log_to_str(
-        self,
-        intermediate_steps: List[Tuple[Any, str]],
-        observation_prefix: str = "Observation: ",
-        llm_prefix: str = "",
-    ) -> str:
         pass
 
     @abstractmethod
